Day 25: Clean up the monad stack a little.
The Beep layer now uses RWST instead of WriterT: the State component
replaces not-so-modularly reaching into the Exec layer's ST, and the
Reader component... is not hugely impactful, but it's there, so why
not store the program in it.

(Note that if Beep did need more than functional state, like for an
MVector, it could always use liftPrim instead of composing simple lifts
like I was doing.)

Also, more type annotations and abbreviations.  Nothing that GHC
actually needs (cf. the rank-2 type on Exec.launch, which *is*
load-bearing), but useful for conceptual clarity.